fixed warping

+++ pulse2percept/electrode2currentmap.py
@@ -889,7 +889,7 @@ class Movie2Pulsetrain(TimeSeries):
                 ppt = np.concatenate((ppt, interpulsegap), axis=0)
                 ppt = np.concatenate((ppt, pulse), axis=0)
 
-        ppt = ppt[0:round(dur / tsample)]
+        ppt = ppt[0:int(round(dur / tsample))]
         intfunc = interpolate.interp1d(np.linspace(0, len(rflum), len(rflum)),
                                        rflum)
